# api/rest/swagger.yaml
openapi: 3.0.3
info:
  title: Fitness Platform API
  version: 1.0.0
  description: API for personalized fitness management with biometric data integration
servers:
  - url: https://fittpulse.ru/api/v1
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /api/v1/training/generate:
    post:
      tags:
        - Training
      summary: Generate training plan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                duration_weeks:
                  type: integer
                  minimum: 1
                  maximum: 52
                available_days:
                  type: array
                  items:
                    type: integer
                    minimum: 0
                    maximum: 6
                class:
                  type: string
                  enum: [strength, cardio, flexibility, hiit, endurance, mixed]
                confidence:
                  type: number
                  minimum: 0
                  maximum: 1
      responses:
        "200":
          description: Plan generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  plan_id:
                    type: string
                  plan_data:
                    type: object
                  training_type:
                    type: string
  /api/v1/training/plans:
    get:
      tags:
        - Training
      summary: Get user training plans
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
      responses:
        "200":
          description: Training plans
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  plans:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
  /api/v1/training/plans/{plan_id}:
    parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Training
      summary: Get specific training plan
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Training plan details
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  plan_id:
                    type: string
                  plan_data:
                    type: object
  /api/v1/training/progress:
    get:
      tags:
        - Training
      summary: Get user training progress
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  progress:
                    type: object
  /api/v1/training/complete:
    post:
      tags:
        - Training
      summary: Complete a workout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - plan_id
                - workout_id
                - rating
              properties:
                plan_id:
                  type: string
                workout_id:
                  type: string
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                feedback:
                  type: string
      responses:
        "200":
          description: Workout completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string

  # ============================================================================
  # PROTECTED ACHIEVEMENTS
  # ============================================================================
  /api/v1/achievements:
    get:
      tags:
        - Achievements
      summary: Get user achievements
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User achievements
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  achievements:
                    type: array
                    items:
                      type: object

  # ============================================================================
  # PROTECTED ML
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    criticalSession:
      type: apiKey
      in: header
      name: X-Critical-Session-Token
...